requirements.txt
matplotlib==3.10.3
pandas==2.3.1
scipy==1.16.0
torch-geometric==2.6.1
einops==0.8.1
fastdtw
torchdiffeq